lambda_mt5_test 2/python/requirements.txt
appdirs==1.4.4
boto3==1.28.62
botocore==1.31.65
numpy==1.21.4
pandas==2.2.0
python-dotenv==1.0.1
Requests==2.31.0
requests_html==0.10.0
schedule==1.2.1
yfinance==0.2.31
